Fix issue #1722: Build warning issues on FreeBSD.
- Using long for break span to solve an overflow as it is to small to handle the MAX SPAN value.
- Using JMP_BUF instead of the plain jmp_buf to solve incompatible pointer assignment.
- hash key can t be null thus the checking is always true.

+++ src/debugger/handlers.h
@@ -96,8 +96,8 @@ struct _xdebug_con {
 #define XDEBUG_RESOLVED_SPAN_MAX  4294967295
 
 struct _xdebug_brk_span {
-	int start;
-	int end;
+	long start;
+	long end;
 };
 
 struct _xdebug_brk_info {
